functions/mr.prope/requirements.txt
azure-functions==1.7.0
numpy==1.20.2
pandas==1.2.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
